Subpackage structure refactored to fit the architecture sketch
