+++ ls-mandelbrot-core/src/main/java/de/lexasoft/mandelbrot/ctrl/package-info.java
@@ -3,8 +3,8 @@
  * <p>
  * In most cases read form a corresponding yaml file.
  * 
- * @see de.lexasoft.mandelbrot.dto.DTO2PropertiesMapper
+ * @see de.lexasoft.mandelbrot.ctrl.DTO2PropertiesMapper
  * @since 1.0
  * @author nierax
  */
-package de.lexasoft.mandelbrot.dto;
+package de.lexasoft.mandelbrot.ctrl;
