+++ ls-mandelbrot-core/src/test/java/de/lexasoft/mandelbrot/MandelbrotCalculationPropertiesTest.java
@@ -21,6 +21,8 @@ import org.junit.jupiter.params.provider.MethodSource;
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.databind.JsonMappingException;
 
+import de.lexasoft.mandelbrot.api.MandelbrotCalculationProperties;
+
 /**
  * @author nierax
  *
