+++ ls-mandelbrot-core/src/test/java/de/lexasoft/mandelbrot/api/MandelbrotRunnerFactoryTest.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package de.lexasoft.mandelbrot;
+package de.lexasoft.mandelbrot.api;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -12,6 +12,14 @@ import java.util.List;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import de.lexasoft.mandelbrot.MandelbrotPointPosition;
+import de.lexasoft.mandelbrot.MandelbrotRunnerFactory;
+import de.lexasoft.mandelbrot.PaletteVariant;
+import de.lexasoft.mandelbrot.api.MandelbrotCalculationProperties;
+import de.lexasoft.mandelbrot.api.MandelbrotRunner;
+import de.lexasoft.mandelbrot.api.MandelbrotRunnerChain;
+import de.lexasoft.mandelbrot.api.MandelbrotSingleRunner;
 
 /**
  * @author nierax
@@ -59,7 +67,7 @@ class MandelbrotRunnerFactoryTest {
 
 	/**
 	 * Test method for
-	 * {@link de.lexasoft.mandelbrot.MandelbrotRunnerFactory#createRunner(de.lexasoft.mandelbrot.MandelbrotCalculationProperties)}.
+	 * {@link de.lexasoft.mandelbrot.MandelbrotRunnerFactory#createRunner(de.lexasoft.mandelbrot.api.MandelbrotCalculationProperties)}.
 	 */
 	@Test
 	void testCreateRunnerMandelbrotCalculationProperties() {
